Change intercity connection check + adding to [Scotland]
